pyproject.toml
[project]
name = "multimodal-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.136.1",
    "fpdf2>=2.8.7",
    "google-genai>=1.73.1",
    "langsmith>=0.7.37",
    "markdown>=3.10.2",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "reportlab>=4.4.10",
    "uvicorn>=0.46.0",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = [
    "httpx>=0.27",
    "pytest>=9.0.3",
]
